requirements.txt
azure-ai-formrecognizer
python-dotenv
pymongo 
dnspython 
botbuilder-core
botbuilder-integration-aiohttp
azure-ai-textanalytics
azure-cosmos
streamlit
azure-storage-blob
pymongo


